Suppress useless-cast warnings from span.hpp on gcc and clang
